# api/public/openapi.yaml
openapi: 3.0.1
info:
  title: Generative Manim API
  description: API for generating Manim code, converting it to video, and handling chat-based code generation.
  version: 1.0.0
  license:
    name: Apache License, Version 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
paths:
...
  /v1/video/rendering:
    post:
      summary: Convert Code to Video
      description: Converts provided Manim code to a video and uploads it to storage.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  description: The Manim code to convert to video.
                file_name:
                  type: string
                  description: The name of the file to save the code.
                file_class:
                  type: string
                  description: The class name to render.
                user_id:
                  type: string
                  description: The user ID for the video.
                  default: "anonymous"
                project_name:
                  type: string
                  description: The project name for the video.
                iteration:
                  type: integer
                  description: The iteration number for naming the video.
                aspect_ratio:
                  type: string
                  description: The aspect ratio of the video (16:9, 1:1, or 9:16).
                  default: "16:9"
                stream:
                  type: boolean
                  description: Whether to stream the percentage of animation shown.
                  default: false
      responses:
        '200':
          description: Successful response with video URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success message.
                  video_url:
                    type: string
                    description: URL of the generated video.
        '207':
          description: Streaming response with animation progress
          content:
            text/event-stream:
              schema:
                type: string
                description: JSON-encoded progress updates or error messages.
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message.
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message.
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - ApiKeyAuth: []